Adapted tests

+++ scamlab-server/src/main/java/resource/StartMenuWSResource.java
@@ -14,7 +14,7 @@ import jakarta.inject.Inject;
 @Authenticated
 @WebSocket(path = "/ws/start-menu")
 public class StartMenuWSResource {
-    public record ChatMessage(long numberOfPlayersConnected) {}
+    public record ChatMessage(int numberOfPlayersConnected) {}
 
     @Inject
     Logger logger;
